@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.27.1
-// 	protoc        v3.6.1
+// 	protoc-gen-go v1.35.2
+// 	protoc        v3.21.12
 // source: shipping/shipping.proto
 
 package shipping
@@ -30,11 +30,9 @@ type CreateShippingRequest struct {
 
 func (x *CreateShippingRequest) Reset() {
 	*x = CreateShippingRequest{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_shipping_shipping_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_shipping_shipping_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CreateShippingRequest) String() string {
@@ -45,7 +43,7 @@ func (*CreateShippingRequest) ProtoMessage() {}
 
 func (x *CreateShippingRequest) ProtoReflect() protoreflect.Message {
 	mi := &file_shipping_shipping_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -75,11 +73,9 @@ type CreateShippingResponse struct {
 
 func (x *CreateShippingResponse) Reset() {
 	*x = CreateShippingResponse{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_shipping_shipping_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_shipping_shipping_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *CreateShippingResponse) String() string {
@@ -90,7 +86,7 @@ func (*CreateShippingResponse) ProtoMessage() {}
 
 func (x *CreateShippingResponse) ProtoReflect() protoreflect.Message {
 	mi := &file_shipping_shipping_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -138,7 +134,7 @@ func file_shipping_shipping_proto_rawDescGZIP() []byte {
 }
 
 var file_shipping_shipping_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_shipping_shipping_proto_goTypes = []interface{}{
+var file_shipping_shipping_proto_goTypes = []any{
 	(*CreateShippingRequest)(nil),  // 0: CreateShippingRequest
 	(*CreateShippingResponse)(nil), // 1: CreateShippingResponse
 }
@@ -156,32 +152,6 @@ func init() { file_shipping_shipping_proto_init() }
 func file_shipping_shipping_proto_init() {
 	if File_shipping_shipping_proto != nil {
 		return
-	}
-	if !protoimpl.UnsafeEnabled {
-		file_shipping_shipping_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CreateShippingRequest); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_shipping_shipping_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*CreateShippingResponse); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
