@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.6.1
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v3.21.12
 // source: order/order.proto
 
 package order
@@ -15,8 +15,13 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	Order_Create_FullMethodName = "/Order/Create"
+	Order_Get_FullMethodName    = "/Order/Get"
+)
 
 // OrderClient is the client API for Order service.
 //
@@ -35,8 +40,9 @@ func NewOrderClient(cc grpc.ClientConnInterface) OrderClient {
 }
 
 func (c *orderClient) Create(ctx context.Context, in *CreateOrderRequest, opts ...grpc.CallOption) (*CreateOrderResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CreateOrderResponse)
-	err := c.cc.Invoke(ctx, "/Order/Create", in, out, opts...)
+	err := c.cc.Invoke(ctx, Order_Create_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -44,8 +50,9 @@ func (c *orderClient) Create(ctx context.Context, in *CreateOrderRequest, opts .
 }
 
 func (c *orderClient) Get(ctx context.Context, in *GetOrderRequest, opts ...grpc.CallOption) (*GetOrderResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(GetOrderResponse)
-	err := c.cc.Invoke(ctx, "/Order/Get", in, out, opts...)
+	err := c.cc.Invoke(ctx, Order_Get_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -54,16 +61,19 @@ func (c *orderClient) Get(ctx context.Context, in *GetOrderRequest, opts ...grpc
 
 // OrderServer is the server API for Order service.
 // All implementations must embed UnimplementedOrderServer
-// for forward compatibility
+// for forward compatibility.
 type OrderServer interface {
 	Create(context.Context, *CreateOrderRequest) (*CreateOrderResponse, error)
 	Get(context.Context, *GetOrderRequest) (*GetOrderResponse, error)
 	mustEmbedUnimplementedOrderServer()
 }
 
-// UnimplementedOrderServer must be embedded to have forward compatible implementations.
-type UnimplementedOrderServer struct {
-}
+// UnimplementedOrderServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedOrderServer struct{}
 
 func (UnimplementedOrderServer) Create(context.Context, *CreateOrderRequest) (*CreateOrderResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Create not implemented")
@@ -72,6 +82,7 @@ func (UnimplementedOrderServer) Get(context.Context, *GetOrderRequest) (*GetOrde
 	return nil, status.Errorf(codes.Unimplemented, "method Get not implemented")
 }
 func (UnimplementedOrderServer) mustEmbedUnimplementedOrderServer() {}
+func (UnimplementedOrderServer) testEmbeddedByValue()               {}
 
 // UnsafeOrderServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to OrderServer will
@@ -81,6 +92,13 @@ type UnsafeOrderServer interface {
 }
 
 func RegisterOrderServer(s grpc.ServiceRegistrar, srv OrderServer) {
+	// If the following call pancis, it indicates UnimplementedOrderServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Order_ServiceDesc, srv)
 }
 
@@ -94,7 +112,7 @@ func _Order_Create_Handler(srv interface{}, ctx context.Context, dec func(interf
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/Order/Create",
+		FullMethod: Order_Create_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(OrderServer).Create(ctx, req.(*CreateOrderRequest))
@@ -112,7 +130,7 @@ func _Order_Get_Handler(srv interface{}, ctx context.Context, dec func(interface
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/Order/Get",
+		FullMethod: Order_Get_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(OrderServer).Get(ctx, req.(*GetOrderRequest))
