@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.6.1
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v3.21.12
 // source: payment/payment.proto
 
 package payment
@@ -15,8 +15,12 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	Payment_Create_FullMethodName = "/Payment/Create"
+)
 
 // PaymentClient is the client API for Payment service.
 //
@@ -34,8 +38,9 @@ func NewPaymentClient(cc grpc.ClientConnInterface) PaymentClient {
 }
 
 func (c *paymentClient) Create(ctx context.Context, in *CreatePaymentRequest, opts ...grpc.CallOption) (*CreatePaymentResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CreatePaymentResponse)
-	err := c.cc.Invoke(ctx, "/Payment/Create", in, out, opts...)
+	err := c.cc.Invoke(ctx, Payment_Create_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -44,20 +49,24 @@ func (c *paymentClient) Create(ctx context.Context, in *CreatePaymentRequest, op
 
 // PaymentServer is the server API for Payment service.
 // All implementations must embed UnimplementedPaymentServer
-// for forward compatibility
+// for forward compatibility.
 type PaymentServer interface {
 	Create(context.Context, *CreatePaymentRequest) (*CreatePaymentResponse, error)
 	mustEmbedUnimplementedPaymentServer()
 }
 
-// UnimplementedPaymentServer must be embedded to have forward compatible implementations.
-type UnimplementedPaymentServer struct {
-}
+// UnimplementedPaymentServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedPaymentServer struct{}
 
 func (UnimplementedPaymentServer) Create(context.Context, *CreatePaymentRequest) (*CreatePaymentResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Create not implemented")
 }
 func (UnimplementedPaymentServer) mustEmbedUnimplementedPaymentServer() {}
+func (UnimplementedPaymentServer) testEmbeddedByValue()                 {}
 
 // UnsafePaymentServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to PaymentServer will
@@ -67,6 +76,13 @@ type UnsafePaymentServer interface {
 }
 
 func RegisterPaymentServer(s grpc.ServiceRegistrar, srv PaymentServer) {
+	// If the following call pancis, it indicates UnimplementedPaymentServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Payment_ServiceDesc, srv)
 }
 
@@ -80,7 +96,7 @@ func _Payment_Create_Handler(srv interface{}, ctx context.Context, dec func(inte
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/Payment/Create",
+		FullMethod: Payment_Create_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(PaymentServer).Create(ctx, req.(*CreatePaymentRequest))
