@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.2.0
-// - protoc             v3.6.1
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v3.21.12
 // source: shipping/shipping.proto
 
 package shipping
@@ -15,8 +15,12 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
+
+const (
+	Shipping_Create_FullMethodName = "/Shipping/Create"
+)
 
 // ShippingClient is the client API for Shipping service.
 //
@@ -34,8 +38,9 @@ func NewShippingClient(cc grpc.ClientConnInterface) ShippingClient {
 }
 
 func (c *shippingClient) Create(ctx context.Context, in *CreateShippingRequest, opts ...grpc.CallOption) (*CreateShippingResponse, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CreateShippingResponse)
-	err := c.cc.Invoke(ctx, "/Shipping/Create", in, out, opts...)
+	err := c.cc.Invoke(ctx, Shipping_Create_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -44,20 +49,24 @@ func (c *shippingClient) Create(ctx context.Context, in *CreateShippingRequest, 
 
 // ShippingServer is the server API for Shipping service.
 // All implementations must embed UnimplementedShippingServer
-// for forward compatibility
+// for forward compatibility.
 type ShippingServer interface {
 	Create(context.Context, *CreateShippingRequest) (*CreateShippingResponse, error)
 	mustEmbedUnimplementedShippingServer()
 }
 
-// UnimplementedShippingServer must be embedded to have forward compatible implementations.
-type UnimplementedShippingServer struct {
-}
+// UnimplementedShippingServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedShippingServer struct{}
 
 func (UnimplementedShippingServer) Create(context.Context, *CreateShippingRequest) (*CreateShippingResponse, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method Create not implemented")
 }
 func (UnimplementedShippingServer) mustEmbedUnimplementedShippingServer() {}
+func (UnimplementedShippingServer) testEmbeddedByValue()                  {}
 
 // UnsafeShippingServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to ShippingServer will
@@ -67,6 +76,13 @@ type UnsafeShippingServer interface {
 }
 
 func RegisterShippingServer(s grpc.ServiceRegistrar, srv ShippingServer) {
+	// If the following call pancis, it indicates UnimplementedShippingServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Shipping_ServiceDesc, srv)
 }
 
@@ -80,7 +96,7 @@ func _Shipping_Create_Handler(srv interface{}, ctx context.Context, dec func(int
 	}
 	info := &grpc.UnaryServerInfo{
 		Server:     srv,
-		FullMethod: "/Shipping/Create",
+		FullMethod: Shipping_Create_FullMethodName,
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(ShippingServer).Create(ctx, req.(*CreateShippingRequest))
